feat(auth): add global auth token provider
