+++ packages/plugins/@nocobase/plugin-auth/src/client/AuthProvider.tsx
@@ -1,41 +1,19 @@
-import {
-  OptionsComponentProvider,
-  SettingsCenterProvider,
-  SigninPageProvider,
-  SignupPageProvider,
-} from '@nocobase/client';
-import React, { FC } from 'react';
-import { Authenticator } from './settings/Authenticator';
-import SigninPage from './basic/SigninPage';
-import { presetAuthType } from '../preset';
-import SignupPage from './basic/SignupPage';
-import { useAuthTranslation } from './locale';
-import { Options } from './basic/Options';
+import { useAPIClient } from '@nocobase/client';
+import React, { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 
-export const AuthProvider: FC = (props) => {
-  const { t } = useAuthTranslation();
-  return (
-    <SettingsCenterProvider
-      settings={{
-        auth: {
-          title: t('Authentication'),
-          icon: 'LoginOutlined',
-          tabs: {
-            authenticators: {
-              title: t('Authenticators'),
-              component: () => <Authenticator />,
-            },
-          },
-        },
-      }}
-    >
-      <OptionsComponentProvider authType={presetAuthType} component={Options}>
-        <SigninPageProvider authType={presetAuthType} tabTitle={t('Sign in via password')} component={SigninPage}>
-          <SignupPageProvider authType={presetAuthType} component={SignupPage}>
-            {props.children}
-          </SignupPageProvider>
-        </SigninPageProvider>
-      </OptionsComponentProvider>
-    </SettingsCenterProvider>
-  );
+export const AuthProvider: React.FC = (props) => {
+  const location = useLocation();
+  const api = useAPIClient();
+
+  useEffect(() => {
+    const params = new URLSearchParams(location.search);
+    const authenticator = params.get('authenticator');
+    const token = params.get('token');
+    if (token) {
+      api.auth.setToken(token);
+      api.auth.setAuthenticator(authenticator);
+    }
+  });
+  return <>{props.children}</>;
 };
