+++ packages/plugins/@nocobase/plugin-auth/src/client/index.tsx
@@ -1,9 +1,11 @@
 import { Plugin } from '@nocobase/client';
+import { AuthPluginProvider } from './AuthPluginProvider';
 import { AuthProvider } from './AuthProvider';
 
 export class AuthPlugin extends Plugin {
   async load() {
-    this.app.use(AuthProvider);
+    this.app.providers.unshift([AuthProvider, {}]);
+    this.app.use(AuthPluginProvider);
   }
 }
