+++ packages/plugins/@nocobase/plugin-oidc/src/server/actions/redirect.ts
@@ -19,7 +19,7 @@ export const redirect = async (ctx: Context, next: Next) => {
   const auth = (await ctx.app.authManager.get(authenticator, ctx)) as OIDCAuth;
   try {
     const { token } = await auth.signIn();
-    ctx.redirect(`${prefix}/signin?authenticator=${authenticator}&token=${token}`);
+    ctx.redirect(`${prefix}/admin?authenticator=${authenticator}&token=${token}`);
   } catch (error) {
     ctx.redirect(`${prefix}/signin?authenticator=${authenticator}&error=${error.message}`);
   }
