+++ packages/plugins/@nocobase/plugin-saml/src/client/SAMLButton.tsx
@@ -1,5 +1,5 @@
 import { LoginOutlined } from '@ant-design/icons';
-import { Authenticator, css, useAPIClient, useCurrentUserContext, useRedirect } from '@nocobase/client';
+import { Authenticator, css, useAPIClient } from '@nocobase/client';
 import { Button, Space, message } from 'antd';
 import React, { useEffect } from 'react';
 import { useSamlTranslation } from './locale';
@@ -8,9 +8,7 @@ import { useLocation } from 'react-router-dom';
 export const SAMLButton = ({ authenticator }: { authenticator: Authenticator }) => {
   const { t } = useSamlTranslation();
   const api = useAPIClient();
-  const redirect = useRedirect();
   const location = useLocation();
-  const { refreshAsync: refresh } = useCurrentUserContext();
 
   const login = async () => {
     const response = await api.request({
@@ -27,7 +25,6 @@ export const SAMLButton = ({ authenticator }: { authenticator: Authenticator }) 
 
   useEffect(() => {
     const params = new URLSearchParams(location.search);
-    const token = params.get('token');
     const name = params.get('authenticator');
     const error = params.get('error');
     if (name !== authenticator.name) {
@@ -35,14 +32,6 @@ export const SAMLButton = ({ authenticator }: { authenticator: Authenticator }) 
     }
     if (error) {
       message.error(error);
-      return;
-    }
-    if (token) {
-      api.auth.setToken(token);
-      api.auth.setAuthenticator(name);
-      refresh()
-        .then(() => redirect())
-        .catch((err) => console.log(err));
       return;
     }
   });
